Added test for Lightfoot.
The benefit of the generic test approach is that the test itself is only defined once. This benefit will be very apparent as new features are added and require testing for each race. Each race must also define the expected properties or their tests will fail, so it also enforces that all of the data for each race is tested.